[SBS-12]: move slideshow to folder components
