+++ app/components/SlideShow.tsx
